A few Pep8 changes and moved Global lists to Dicts
Returned and indented some of the long Global Constants down to Pep8
standards. Didn't touch the actual code itself but possibly something to
do in the future.

Moved all of the global arrays into team specific Dictionaries. This is
a slightly better way of doing this I think because you don't need to
run the build_arrays method at all, you can just reference the item you
want and spit it out that way.

I haven't really been able to test this code as there are not games
currently to pull from. I'll look at that when the preseason kicks off.